Unit tests pass Still need to get to :100: coverage
