+++ client/src/test/suite/extension.test.ts
@@ -923,224 +923,6 @@ suite('Extension Test Suite', function (this: Suite) {
     const expected = ['onDidChangeWorkspaceFolders', 'test_event']
     assert((mockClient.sendRequest as SinonRoot.SinonStub).calledOnceWith(...expected))
   })
-  test('TocTreesProvider returns expected TocTreeItems', async () => {
-    const fakeTreeCollection: BookToc[] = []
-    fakeTreeCollection.push(
-      {
-        type: BookRootNode.Singleton,
-        title: 'Collection1',
-        slug: 'collection1',
-        uuid: '',
-        language: '',
-        licenseUrl: '',
-        absPath: 'path/to/nowhere-book',
-        tree: [{
-          type: TocNodeKind.Inner,
-          value: { token: 'id123', title: 'subcollection' },
-          children: [{
-            type: TocNodeKind.Leaf,
-            value: {
-              token: 'id234',
-              absPath: 'path/to/nowhere',
-              fileId: 'm00001',
-              title: 'Module1'
-            }
-          },
-          {
-            type: TocNodeKind.Leaf,
-            value: {
-              token: 'id345',
-              absPath: 'path/to/nowhere2',
-              fileId: 'm00002',
-              title: 'Module2'
-            }
-          }]
-        }]
-      },
-      {
-        type: BookRootNode.Singleton,
-        title: 'Collection2',
-        slug: 'collection2',
-        uuid: '',
-        language: '',
-        licenseUrl: '',
-        absPath: 'path/to/nowhere-book',
-        tree: [{
-          type: TocNodeKind.Leaf,
-          value: {
-            token: 'id123',
-            absPath: 'path/to/nowhere3',
-            fileId: 'm00003',
-            title: 'Module3'
-          }
-        }]
-      }
-    )
-    const fakeWorkspacePath = '/tmp/fakeworkspace'
-    sinon.stub(utils, 'getRootPathUri').returns(vscode.Uri.file(fakeWorkspacePath))
-    const module1Item = new TocTreeItem(
-      TocItemIcon.Page,
-      'Module1',
-      vscode.TreeItemCollapsibleState.None,
-      [],
-      {
-        title: 'open',
-        command: 'vscode.open',
-        arguments: [vscode.Uri.file(`${fakeWorkspacePath}/modules/m00001/index.cnxml`)]
-      },
-      'm00001'
-    )
-    const module2Item = new TocTreeItem(
-      TocItemIcon.Page,
-      'Module2',
-      vscode.TreeItemCollapsibleState.None,
-      [],
-      {
-        title: 'open',
-        command: 'vscode.open',
-        arguments: [vscode.Uri.file(`${fakeWorkspacePath}/modules/m00002/index.cnxml`)]
-      },
-      'm00002'
-    )
-    const module3Item = new TocTreeItem(
-      TocItemIcon.Page,
-      'Module3',
-      vscode.TreeItemCollapsibleState.None,
-      [],
-      {
-        title: 'open',
-        command: 'vscode.open',
-        arguments: [vscode.Uri.file(`${fakeWorkspacePath}/modules/m00003/index.cnxml`)]
-      },
-      'm00003'
-    )
-    const module3ItemToggled = new TocTreeItem(
-      TocItemIcon.Page,
-      'Module3 (m00003)',
-      vscode.TreeItemCollapsibleState.None,
-      [],
-      {
-        title: 'open',
-        command: 'vscode.open',
-        arguments: [vscode.Uri.file(`${fakeWorkspacePath}/modules/m00003/index.cnxml`)]
-      },
-      undefined
-    )
-    const subcollectionItem = new TocTreeItem(
-      TocItemIcon.SubBook,
-      'subcollection',
-      vscode.TreeItemCollapsibleState.Collapsed,
-      [module1Item, module2Item]
-    )
-    const collection1Item = new TocTreeItem(
-      TocItemIcon.Book,
-      'Collection1',
-      vscode.TreeItemCollapsibleState.Collapsed,
-      [subcollectionItem],
-      {
-        title: 'open',
-        command: 'vscode.open',
-        arguments: [vscode.Uri.file(`${fakeWorkspacePath}/collections/collection1.collection.xml`)]
-      }
-    )
-    const collection2Item = new TocTreeItem(
-      TocItemIcon.Book,
-      'Collection2',
-      vscode.TreeItemCollapsibleState.Collapsed,
-      [module3Item],
-      {
-        title: 'open',
-        command: 'vscode.open',
-        arguments: [vscode.Uri.file(`${fakeWorkspacePath}/collections/collection2.collection.xml`)]
-      }
-    )
-
-    const mockClient = createMockClient()
-    // We don't want to just return []
-    const sendRequestMock = sinon.stub()
-    mockClient.sendRequest = sendRequestMock
-    const tocTreesProvider = new TocTreesProvider({ bookTocs: DEFAULT_BOOK_TOCS_ARGS, resourceRootDir, client: mockClient, events: createMockEvents().events })
-    sendRequestMock.onCall(0).resolves(null)
-    sendRequestMock.onCall(1).resolves(fakeTreeCollection)
-
-    assert.deepStrictEqual(await tocTreesProvider.getChildren(undefined), [])
-    assert.deepStrictEqual(await tocTreesProvider.getChildren(undefined), [collection1Item, collection2Item])
-    assert.deepStrictEqual(await tocTreesProvider.getChildren(collection2Item), [module3Item])
-    assert.deepStrictEqual(tocTreesProvider.getTreeItem(collection2Item), collection2Item)
-    assert.deepStrictEqual(tocTreesProvider.getParent(collection2Item), undefined)
-    assert.deepStrictEqual(tocTreesProvider.getParent(module3Item), collection2Item)
-    assert.deepStrictEqual(tocTreesProvider.getParent(module1Item), subcollectionItem)
-    tocTreesProvider.toggleFilterMode()
-    assert.deepStrictEqual(tocTreesProvider.getTreeItem(module3Item), module3ItemToggled)
-  })
-  test('toggleTocTreesFilteringHandler', async () => {
-    const revealStub = sinon.stub()
-    const toggleFilterStub = sinon.stub()
-    const getChildrenStub = sinon.stub()
-    const refreshStub = sinon.stub()
-
-    const view: vscode.TreeView<BookOrTocNode> = {
-      reveal: revealStub
-    } as unknown as vscode.TreeView<BookOrTocNode>
-    const provider: TocsTreeProvider = {
-      toggleFilterMode: toggleFilterStub,
-      getChildren: getChildrenStub,
-      refresh: refreshStub
-    } as unknown as TocsTreeProvider
-    const fakeChildren = [
-      { label: 'col1', children: [{ label: 'subcol', children: [{ label: 'm2', children: [] }] }] },
-      { label: 'col2', children: [{ label: 'm1', children: [] }] }
-    ]
-    getChildrenStub.resolves(fakeChildren)
-
-    const handler = toggleTocTreesFilteringHandler(view, provider)
-    await handler()
-    assert(toggleFilterStub.calledOnce)
-    assert(getChildrenStub.calledOnce)
-    assert(revealStub.calledThrice)
-    assert(revealStub.calledWith(fakeChildren[0], { expand: true }))
-    assert(revealStub.calledWith(fakeChildren[0].children[0], { expand: true }))
-    assert(revealStub.calledWith(fakeChildren[1], { expand: true }))
-    assert(refreshStub.notCalled)
-  })
-  test('toggleTocTreesFilteringHandler disables itself while revealing', async () => {
-    const revealStub = sinon.stub()
-    const toggleFilterStub = sinon.stub()
-    const getChildrenStub = sinon.stub()
-    const fakeChildren = [
-      { label: 'col1', children: [{ label: 'm1', children: [] }] }
-    ]
-    getChildrenStub.resolves(fakeChildren)
-
-    const view: vscode.TreeView<BookOrTocNode> = {
-      reveal: revealStub
-    } as unknown as vscode.TreeView<BookOrTocNode>
-    const provider: TocsTreeProvider = {
-      toggleFilterMode: toggleFilterStub,
-      getChildren: getChildrenStub
-    } as unknown as TocsTreeProvider
-
-    const handler = toggleTocTreesFilteringHandler(view, provider)
-    // Invoke the handler the first time reveal is called to simulate a parallel
-    // user request without resorting to synthetic delay injection
-    revealStub.onCall(0).callsFake(handler)
-    await handler()
-    assert(toggleFilterStub.calledOnce)
-    assert(revealStub.calledOnce)
-    assert(getChildrenStub.calledOnce)
-  })
-  test('toggleTocTreesFilteringHandler doesn\'t lock itself on errors', async () => {
-    const toggleFilterStub = sinon.stub().throws()
-    const view: vscode.TreeView<BookOrTocNode> = {} as unknown as vscode.TreeView<BookOrTocNode>
-    const provider: TocsTreeProvider = {
-      toggleFilterMode: toggleFilterStub
-    } as unknown as TocsTreeProvider
-
-    const handler = toggleTocTreesFilteringHandler(view, provider)
-    try { await handler() } catch { }
-    try { await handler() } catch { }
-    assert(toggleFilterStub.calledTwice)
-  })
   test('TocTreesProvider fires event on refresh', async () => {
     const tocTreesProvider = new TocTreesProvider({ bookTocs: DEFAULT_BOOK_TOCS_ARGS, resourceRootDir, client: createMockClient(), events: createMockEvents().events })
     const eventFire = sinon.stub((tocTreesProvider as any)._onDidChangeTreeData, 'fire')
